feat: changes file routes for card component and finished major updates to card
