+++ modules/explore/searchPage.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Card from './components/Card';
+import Card from './components/Card-old';
 import { UserInfo } from './@types';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
